argocd_stage/hit-counter/requirements.txt
Flask
parse
pytest-flask
gunicorn
